Refactor to use API in various commands
