+++ librz/core/cio.c
@@ -17,12 +17,13 @@ RZ_API int rz_core_setup_debugger(RzCore *r, const char *debugbackend, bool atta
 	}
 
 	rz_config_set(r->config, "io.ff", "true");
-	rz_core_cmdf(r, "dL %s", debugbackend);
+	rz_config_set(r->config, "dbg.backend", debugbackend);
 	if (!is_gdb) {
 		pid = rz_io_desc_get_pid(fd);
-		rz_core_cmdf(r, "dp=%d", pid);
+		rz_debug_select(r->dbg, pid, r->dbg->tid);
+		r->dbg->main_pid = pid;
 		if (attach) {
-			rz_core_cmdf(r, "dpa %d", pid);
+			rz_core_debug_attach(r, pid);
 		}
 	}
 	//this makes to attach twice showing warnings in the output
